+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".main_activity.MainActivity">
+        <activity android:name=".ui.main_activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
