+++ app/src/main/java/com/ehealth4everyone/olamideadeleye/di/CarOwnerComponent.java
@@ -1,6 +1,6 @@
 package com.ehealth4everyone.olamideadeleye.di;
 
-import com.ehealth4everyone.olamideadeleye.car_owners_fragment.CarOwnerFragment;
+import com.ehealth4everyone.olamideadeleye.ui.car_owners_fragment.CarOwnerFragment;
 
 import dagger.Subcomponent;
