+++ app/src/main/java/com/ehealth4everyone/olamideadeleye/repo/FilterRepo.java
@@ -1,8 +1,6 @@
 package com.ehealth4everyone.olamideadeleye.repo;
 
-import android.content.Context;
-
-import com.ehealth4everyone.olamideadeleye.filter_model.Filter;
+import com.ehealth4everyone.olamideadeleye.models.Filter;
 
 import java.util.List;
