+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main_activity.MainActivity">
+    tools:context=".ui.main_activity.MainActivity">
 
     <ProgressBar
         android:id="@+id/progress_circular"
